chore(CartList): Remove component - depracated
